Workaround ChangelogOpen race
It appears that RegisterUser is enabling changelogs
asyncronously, and sometimes when we call ChangelogOpen() immediately
afterwards changelogs have not been enabled yet.  Unfortunately, the
llapi_changelog_start() call doesn't seem to detect this, and returns
a broken changlog handle.  This appears to be why we occasionaly see
a series of "Unknown changelog message type 0:0" errors and a failed
test.

The workaround is to open the changelog each time we attempt to read
it.