Refactor changelog code based on feedback
Provides a MDT-specific changelog.Handle which can be used to
get changelog records and clear old records.

Also provides a standalone changelog follower as a workaround for
the fact that the follow flag doesn't work in liblustreapi. This
may or may not be useful, and we can hopefully get rid of it
altogether if liblustreapi gets fixed.