Pass RootDir to hsm functions instead of fs ID
The fs ID is a unique identifier for a filesystem that is used
to identify a filesystem independent of a particular mountpoint.

The RootDir is a specific mount point of an fs ID. Since we want
the HSM commands to execute on a specific mount point, we should
be passing RootDir to those functions. The patch updates that.

For convenience, this also adds a Root() method for ID, however
once we actually implement ID properly, this might not actually be
doable.

RootDir is probably the best name.