Convert all offset/length values to int64
The POSIX standard specifies that these values should be represented
by a signed integer, and the Go stdlib adheres to this standard. By
changing our code to convert between uint64 and int64 at the llapi
boundary, we can simplify the rest of our code and also have more
confidence that we'll catch any overflow issues.